+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to
 
 - @cosmjs/cosmwasm-stargate: Export types `Code`, `CodeDetails`, `Contract`,
   `ContractCodeHistoryEntry` and `JsonObject` which are response types of
-  `CosmWasmClient` methods.
+  `CosmWasmClient` methods. Export types `ChangeAdminResult`, `ExecuteResult`,
+  `InstantiateOptions`, `InstantiateResult`, `MigrateResult`, `UploadMeta` and
+  `UploadResult` which are argument or response types of `SigningCosmWasmClient`
+  methods.
 
 ### Fixed